requirements.txt
-e .
aws-cdk.aws-s3
black
